adds compilation support as java 9 module; move tests to external module 'ssl-client-tests';
