--- src/main/java/io/net/ssl/client/ClientSslProperties.java
@@ -1,20 +1,0 @@
-package io.net.ssl.client;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Data
-public class ClientSslProperties {
-
-    private String keyStore;
-    private String keyStorePassword;
-    private String keyStoreType;
-    private String keyStoreProvider;
-
-    private boolean disableSslValidation = false;
-    private String trustStore;
-    private String trustStorePassword;
-    private String trustStoreType;
-    private String trustStoreProvider;
-
-}
