--- src/main/java/io/netty/handler/ssl/NettySslContextBuilderFactory.java
@@ -1,18 +1,0 @@
-package io.netty.handler.ssl;
-
-import lombok.RequiredArgsConstructor;
-
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.TrustManagerFactory;
-
-import static io.netty.handler.ssl.SslContextBuilder.forClient;
-
-@RequiredArgsConstructor
-public class NettySslContextBuilderFactory {
-    private final KeyManagerFactory keyManagerFactory;
-    private final TrustManagerFactory trustManagerFactory;
-
-    public SslContextBuilder newSslContextBuilder() {
-        return forClient().trustManager(trustManagerFactory).keyManager(keyManagerFactory);
-    }
-}
