+++ src/main/java/m4gshm/springframework/ssl/client/netty/handler/NettySslContextBuilderFactory.java
@@ -1,0 +1,19 @@
+package m4gshm.springframework.ssl.client.netty.handler;
+
+import io.netty.handler.ssl.SslContextBuilder;
+import lombok.RequiredArgsConstructor;
+
+import javax.net.ssl.KeyManagerFactory;
+import javax.net.ssl.TrustManagerFactory;
+
+import static io.netty.handler.ssl.SslContextBuilder.forClient;
+
+@RequiredArgsConstructor
+public class NettySslContextBuilderFactory {
+    private final KeyManagerFactory keyManagerFactory;
+    private final TrustManagerFactory trustManagerFactory;
+
+    public SslContextBuilder newSslContextBuilder() {
+        return forClient().trustManager(trustManagerFactory).keyManager(keyManagerFactory);
+    }
+}
