+++ src/main/java/m4gshm/springframework/ssl/client/utils/AcceptAnyHostname.java
@@ -1,0 +1,11 @@
+package m4gshm.springframework.ssl.client.utils;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.SSLSession;
+
+public class AcceptAnyHostname implements HostnameVerifier {
+    @Override
+    public boolean verify(String hostname, SSLSession session) {
+        return true;
+    }
+}
