--- src/main/java/org/springframework/http/client/SslVerificationHttpRequestFactory.java
@@ -1,30 +1,0 @@
-package org.springframework.http.client;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import java.io.IOException;
-import java.net.HttpURLConnection;
-
-@Slf4j
-@RequiredArgsConstructor
-public class SslVerificationHttpRequestFactory extends SimpleClientHttpRequestFactory {
-
-    final protected SSLContext sslContext;
-    final protected HostnameVerifier hostnameVerifier;
-
-    @Override
-    protected void prepareConnection(HttpURLConnection connection, String httpMethod) throws IOException {
-        if (connection instanceof HttpsURLConnection) prepareHttpsConnection((HttpsURLConnection) connection);
-        super.prepareConnection(connection, httpMethod);
-    }
-
-    protected void prepareHttpsConnection(HttpsURLConnection connection) {
-        connection.setSSLSocketFactory(sslContext.getSocketFactory());
-        if (hostnameVerifier != null) connection.setHostnameVerifier(hostnameVerifier);
-    }
-
-}
