+++ ssl-client-tests/src/test/resources/generate-ssl-keystore.sh
